requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-pinecone
simsimd
langgraph
pinecone
beautifulsoup4
python-dotenv
pyTelegramBotAPI
pydantic
